requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url --no-emit-trusted-host --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   groq
    #   httpx
    #   jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.2.0
    # via
    #   jsonlines
    #   jsonschema
    #   referencing
babel==2.14.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via
    #   ipython
    #   tqdm
comm==0.2.1
    # via ipykernel
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   retry
defusedxml==0.7.1
    # via nbconvert
distro==1.9.0
    # via groq
executing==2.0.1
    # via stack-data
fastjsonschema==2.19.1
    # via nbformat
fonttools==4.49.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
google-api-core==2.18.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.16.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-storage
googleapis-common-protos==1.63.0
    # via google-api-core
groq==0.4.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via
    #   groq
    #   jupyterlab
idna==3.6
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.29.2
    # via jupyterlab
ipython==8.22.1
    # via
    #   -r requirements.in
    #   ipykernel
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.9.17
    # via jupyterlab-server
jsonlines==4.0.0
    # via -r requirements.in
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.21.1
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.9.0
    # via jupyter-server
jupyter-lsp==2.2.2
    # via jupyterlab
jupyter-server==2.12.5
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.2
    # via jupyter-server
jupyterlab==4.1.2
    # via notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.25.3
    # via
    #   jupyterlab
    #   notebook
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.8.3
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
nbclient==0.9.0
    # via nbconvert
nbconvert==7.16.1
    # via jupyter-server
nbformat==5.9.2
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.1.0
    # via -r requirements.in
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
overrides==7.7.0
    # via jupyter-server
packaging==23.2
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
pandas==2.2.1
    # via
    #   -r requirements.in
    #   seaborn
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.3
    # via jedi
pillow==10.2.0
    # via matplotlib
platformdirs==4.2.0
    # via jupyter-core
prometheus-client==0.20.0
    # via jupyter-server
prompt-toolkit==3.0.43
    # via ipython
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==5.9.8
    # via ipykernel
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via retry
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==2.6.4
    # via
    #   -r requirements.in
    #   groq
    #   pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pygments==2.17.2
    # via
    #   ipython
    #   nbconvert
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via pydantic-settings
python-json-logger==2.0.7
    # via jupyter-events
pytz==2024.1
    # via pandas
pywin32==306
    # via jupyter-core
pywinpty==2.0.12
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   terminado
pyyaml==6.0.1
    # via jupyter-events
pyzmq==25.1.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
    #   jupyterlab-server
retry==0.9.2
    # via -r requirements.in
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
seaborn==0.13.2
    # via -r requirements.in
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via
    #   anyio
    #   groq
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.0
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.66.2
    # via -r requirements.in
traitlets==5.14.1
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.8.19.20240106
    # via arrow
typing-extensions==4.10.0
    # via
    #   groq
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.7.0
    # via jupyter-server
